_src/requirements.txt
requests
beautifulsoup4
GitPython